+++ src/app/pages/course/course.component.ts
@@ -1,6 +1,6 @@
-import { Component, inject, ViewChild } from '@angular/core';
+import { Component, ViewChild } from '@angular/core';
 import { CourseService } from '../../services/course.service';
-import { Course} from '../../model/course';
+import { Course } from '../../model/course';
 import { MatTableDataSource, MatTableModule } from '@angular/material/table';
 import { MatIconModule } from '@angular/material/icon';
 import { MatFormFieldModule } from '@angular/material/form-field';
@@ -8,15 +8,18 @@ import { MatInputModule } from '@angular/material/input';
 import { MatPaginator, MatPaginatorModule } from '@angular/material/paginator';
 import { MatSort, MatSortModule } from '@angular/material/sort';
 import { RouterLink, RouterOutlet } from '@angular/router';
-import {MatSnackBar, MatSnackBarModule} from '@angular/material/snack-bar';
+import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
 import { switchMap } from 'rxjs';
 import { MatButtonModule } from '@angular/material/button';
+import { CommonModule } from '@angular/common';
+import { ConfirmationDeleteComponent } from './confirmation-delete/confirmation-delete.component';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-course',
+  standalone: true,
   imports: [
     MatTableModule,
-    MatIconModule,
     MatIconModule,
     MatFormFieldModule,
     MatInputModule,
@@ -25,20 +28,21 @@ import { MatButtonModule } from '@angular/material/button';
     RouterOutlet,
     RouterLink,
     MatSnackBarModule,
-    MatButtonModule
+    MatButtonModule,
+    CommonModule,
+    MatDialogModule
   ],
   templateUrl: './course.component.html',
   styleUrl: './course.component.css'
 })
 export class CourseComponent {
-  //displayedColumns: string[] = ['idCourse', 'code', 'name'];
   dataSource: MatTableDataSource<Course>;
-  // publishers: Publisher[];
+
   columnsDefinitions = [
-    { def: 'idCourse', label: 'idCourse', hide: true },
-    { def: 'code', label: 'code', hide: false },
-    { def: 'name', label: 'name', hide: false },
-    { def: 'actions', label: 'actions', hide: false },
+    { def: 'idCourse', label: 'ID', hide: true },
+    { def: 'code', label: 'CODE', hide: false },
+    { def: 'name', label: 'NAME', hide: false },
+    { def: 'actions', label: 'ACTIONS', hide: false },
   ];
 
   @ViewChild(MatPaginator) paginator: MatPaginator;
@@ -46,45 +50,60 @@ export class CourseComponent {
 
   constructor(
     private courseService: CourseService,
-    private _snackBar: MatSnackBar
-  ){}
-  //courseService = inject(CourseService);
+    private _snackBar: MatSnackBar,
+    public dialog: MatDialog
+  ) {}
 
   ngOnInit(): void {
-    // this.courseService.findAll().subscribe(data => console.log(data));
-    // this.courseService.findAll().subscribe(data => this.courses = data);
-    this.courseService.findAll().subscribe((data) => {
-      this.createTable(data);
-    });
-
-    this.courseService.getCourseChange().subscribe(data => this.createTable(data))
-  
-   // this._snackBar.open('sample message','INFO', {duration: 2000, horizontalPosition: 'right', verticalPosition:'bottom'});
-   this.courseService.getMessageChange().subscribe(
-    data => this._snackBar.open(data,'INFO', {duration: 2000, horizontalPosition: 'right', verticalPosition:'bottom'})
-   );
+    this.loadCourses();
+    
+    this.courseService.getCourseChange().subscribe(data => this.createTable(data));
+    
+    this.courseService.getMessageChange().subscribe(
+      data => this._snackBar.open(data, 'INFO', { 
+        duration: 2000, 
+        horizontalPosition: 'right', 
+        verticalPosition: 'bottom' 
+      })
+    );
   }
 
-  createTable(data: Course[]){
+  loadCourses() {
+    this.courseService.findAll().subscribe(data => {
+      this.createTable(data);
+    });
+  }
+
+  createTable(data: Course[]) {
     this.dataSource = new MatTableDataSource(data);
     this.dataSource.paginator = this.paginator;
     this.dataSource.sort = this.sort;
   }
 
   getDisplayedColumns() {
-    return this.columnsDefinitions.filter((cd) => !cd.hide).map((cd) => cd.def);
+    return this.columnsDefinitions
+      .filter(cd => !cd.hide)
+      .map(cd => cd.def);
   }
 
   applyFilter(e: any) {
-    this.dataSource.filter = e.target.value.trim();
+    this.dataSource.filter = e.target.value.trim().toLowerCase();
   }
 
-  delete(id: number){
-    this.courseService.delete(id)
-      .pipe(switchMap( () => this.courseService.findAll()))
-      .subscribe( data => {
-        this.courseService.setCourseChange(data);
-        this.courseService.setMessageChange('DELETED!');
-      });
+  delete(id: number, name: string) {
+    const dialogRef = this.dialog.open(ConfirmationDeleteComponent, {
+      data: { name: name }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+        this.courseService.delete(id)
+          .pipe(switchMap(() => this.courseService.findAll()))
+          .subscribe(data => {
+            this.courseService.setCourseChange(data);
+            this.courseService.setMessageChange('Course deleted successfully!');
+          });
+      }
+    });
   }
 }
